+++ doc/changelog.md
@@ -6,11 +6,24 @@ All notable changes to Mili-python will be documented in this file.
 
 ### Added
 
+### Fixed
+
+### Changed
+
+### Removed
+
+## [v0.7.7] - 2024-8-28
+
+### Added
+
 - Updated the functions `AdjacencyMapping.nearest_element` and `AdjacencyMapping.nearest_node` to support filtering by material name/number.
+- Added function `state_variable_titles` to get titles of state variables.
 
 ### Fixed
 
 ### Changed
+
+- Removed numpy 2.0 version limit
 
 ### Removed